CLIENT LIBRARY FOR HaaS
Base code for the client library is ready.
1.) client.py consists of all the library code
2.) Exceptions are defined in error_clientlib.py
3.) cli.py now relies on client.py for its functioning

Note: There was lot of coding that I have done
and it took me lot of time and effor to test calls.

I will improve the feedback from cli and exceptions
based on feedback from fellow reviewers.

I could not test certain code related to headnodes and consoles.
Will find a way to do that later.

This is commit is a basic work towards getting fully functional library.